# Optimize `remove_tracks` by skipping unnecessary iteration