Depends on #1083.

It is difficult to trace the changes as this PR depends on #1083 but the key change is the following:

```diff
     if (is_valid_thread) {
-
         const auto id = sh_meas_ids[threadIndex];
-        is_duplicate = false;
-
-        for (unsigned int i = 0; i < threadIndex; ++i) {
-            if (sh_meas_ids[i] == id) {
-                is_duplicate = true;
-                break;
-            }
-        }
+        is_duplicate = (threadIndex > 0 && sh_meas_ids[threadIndex - 1] == id);
     }
```
This simplification is possible because the inputs to `sh_meas_ids` are already sorted in the previous kernel (`count_removable_tracks`)

The performance is improved by about 10%

Before
```
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from CUDAStandard/CUDAGreedyResolutionCompareToCPU
[ RUN      ] CUDAStandard/CUDAGreedyResolutionCompareToCPU.Comparison/0
Event: 0 Seed: 42
 Time for the cpu method 471 ms
 Time for the cuda method 138 ms
Event: 1 Seed: 43
 Time for the cpu method 491 ms
 Time for the cuda method 97 ms
Event: 2 Seed: 44
 Time for the cpu method 464 ms
 Time for the cuda method 98 ms
Event: 3 Seed: 45
 Time for the cpu method 482 ms
 Time for the cuda method 101 ms
Event: 4 Seed: 46
 Time for the cpu method 482 ms
 Time for the cuda method 100 ms
```

After
```
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from CUDAStandard/CUDAGreedyResolutionCompareToCPU
[ RUN      ] CUDAStandard/CUDAGreedyResolutionCompareToCPU.Comparison/0
Event: 0 Seed: 42
 Time for the cpu method 477 ms
 Time for the cuda method 139 ms
Event: 1 Seed: 43
 Time for the cpu method 466 ms
 Time for the cuda method 88 ms
Event: 2 Seed: 44
 Time for the cpu method 474 ms
 Time for the cuda method 89 ms
Event: 3 Seed: 45
 Time for the cpu method 486 ms
 Time for the cuda method 91 ms
Event: 4 Seed: 46
 Time for the cpu method 495 ms
 Time for the cuda method 91 ms
 ```